FIX: Remove buggy mixing samplerate fallback
This driver wasn’t giving the system a chance to switch frequencies, so
the mixer samplerate and system sound card samplerate could end up
mismatched (“chipmunk effect”). Now we just rely on the driver to
return an error if the frequency is unsupported and don’t try to do any
falling back.