Cocoa: Add drag and drop support
This commit adds drag & drop file loading support to MilkyTracker.
When a valid file is dragged into MilkyTracker's window, a highlighted
border is drawn around the window to indicate a valid drop target.
This border is drawn as a triangle strip, using the system highlight
colour as defined in Settings > General.

I've also optimised the drawing of the main UI quad by removing the
need for an EBO and drawing the quad as a triangle strip instead.