Posix: Follow XDG basedir spec for config dir
http://standards.freedesktop.org/basedir-spec/basedir-spec-latest.html

This closes issue #12.

Milkytracker will check for a config file in the old directory
`$HOME/.milkytracker_config` and if it exists will move it to the new
XDG basedir compliant directory, but not if one already exists - in case
the user wants to use an older version of Milkytracker with a separate
config.