Windows: Fix waveOut crashes and deadlocks
This commit attempts to fix the waveOut driver's worst problems.

With a 64bit build, we were losing precision and crashing because of
DWORD casts which should have been DWORD_PTR. Additionally, deadlock was
being caused at least on Windows 8, because calls to the waveOut family
of functions were present within the waveOut callback itself, which is
not allowed according to MSDN:
https://msdn.microsoft.com/en-us/library/dd743869.aspx ("Calling other
wave functions will cause deadlock")

Now, instead of a callback, a worker thread is spawned to handle
refilling the buffer, and waveOut is set up to pass it a message
to notify it that a block has been consumed. This avoids the above
problem.

I also cannot see a reason for keeping the timeEmulation functionality,
which was the only difference between "WaveOut (old)" and "WaveOut
(Vista)" in MilkyTracker. I'm going to assume this was a hack to
workaround previous problems which may have actually been caused by the
above, so I've removed this, and now we have a single WaveOut option for
all Windows versions.