Add sample offset tool/preview ability
This commit lets you hold down the CTRL (Command on Mac) key in order to
see the nearest sample offset value for the 9xx command.
Additionally, you can hold the left mouse button to preview the sample
from that offset. Only works for the valid range of the 9xx command
(i.e. the first 64kb of the sample).