Cocoa: Fix build for macOS 10.7
We were missing some ARC (automatic reference counting) compile flags on the
subsystems that interface with Cocoa (and therefore contain Objective-C code).

This caused segfaults when trying to run MilkyTracker on macOS Lion.

Enabling ARC also revealed a programming error in DisplayDevice_COCOA.mm which
caused a compile error with this flag enabled.