Cocoa: Add macOS Lion compatibility flags to ibtool
When building using the Makefile generator instead of Xcode, ibtool has to be
invoked manually in our CMake scripts to compile Interface Builder layouts.
Xcode automatically invokes ibtool with the correct flags based on the project
Deployment Target, but we don't - so UI resources compiled with 'make' wouldn't
work on macOS 10.7.

Add the '--minimum-deployment-target 10.7' flag so that when using the Makefile
build method MilkyTracker will work properly on macOS Lion.