Fix noise generator on 64-bit platforms
The cause is the random number generator uses a union to access the
same data structure both as integer and as floating point variable in
src/tracker/VRand.h

However the integer part of the union is defined as "unsigned long" -
unsigned long does not have a guaranteed fixed width on all system, as
such it might or might not fit the "float" in the same union depending
on the systems bit width.

Bug report and patch provided by CorvusCorax, see GitHub issue #79 for
details.