Cocoa: Switch from Homebrew libraries to submodules
Homebrew only provides binaries that are built for the macOS API level of the
machine that Homebrew exists on, which makes Homebrew-provided libraries
unsuitable for linking into apps that will run on older versions of macOS.

To allow building MilkyTracker for older versions of macOS, we must build our
own versions of these libraries with the -mmacosx-version-min flag set to match
the deployment target we are aiming for (in our case 10.7). This compile flag
affects many macros in the Apple SDK and adds metadata to the object files
indicating their compatibility level, so it is important to use it properly.

On macOS, the CMake project now calls the external build systems of these
libraries and passes our CMake-specified CFLAGS along to them.