Fix a compiler error with with libc++ 3.8.0:
TrackerStartUp.cpp:129:12: error: call to 'abs' is ambiguous
                deltaT = abs(::PPGetTickCount() - startTime);
                         ^~~

This is because PPGetTickCount() returns an unsigned value, and it is therefore
ambiguous to which type the expression should be casted before calling abs().
A similar ambiguity exists further down in the cpp file.  Fix these by explicity
casting the PPGetTickCount() return value to a signed type.

This bug was originally reported in the FreeBSD bugtracker:

https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=208489