Full height pattern sequencer
This is an experiment to improve the UI by giving the pattern sequencer more vertical space.
It is also known in the code as the order list section.

The idea comes from Renoise. This change is an unfinished example of what it could look like.
To minimize the code changes, the pattern sequencer is not completely full height and the old
order list container needs cleanup.

The goal of this PR is to discuss if UI changes like these could be desired. Or if MilkyTracker
is aimed at staying a true FT2 clone.