Enable RecPosProvider roundToClosestRow
Ok so according to the blame this code is already in milky for decades,
just hard-coded to be off rofl.

This commit enables the roundToClosestRow bool. While somewhat working
as expected, an annoying bug I encountered is that rounded notes tend to
be played twice.
 - Once in RecorderLogic: `playerLogic->playNote()`
 - Once in the module player, once we arrive at the displaced note
   couple of ticks later.

One option around this bug is to disable playNote() when rounding the
row. This is what I tried to do while fixing the same bug for schism,
but when I checked it out it appeared to feel unnatural. I want my notes
to be played right away, play the quantized version next iteration ffs.

So I took a different approach. Considering I can't eliminate playNote,
I need to prevent the player from actually playing the displaced note.

Eventually, my workaround is to introduce a new "ghost effect", one that
doesn't get rendered in the pattern editor. The new effect (0x80) marks
the cell to be skipped in the player. Once arriving at the player, a bit
of code removes the 0x80 effect, allowing normal play next visit.

Note currently only XM player supports the 0x80 effect. Other players
will probably still play double notes.

GUI setting to control roundToClosestRow coming in the next commit.