Use UTC when writing metainfo timestamp
To help with reproducible builds[1] the `string(TIMESTAMP)` CMake
function respects `$SOURCE_DATE_EPOCH` if set and uses that instead of
the current time. However, the output is still sensitive to the local
timezone which is enough to make the build un-reproducible.

Fix this by using UTC when getting the timestamp.

[1] https://reproducible-builds.org/